Added a public dto for highscores with teamLogotype.
