+++ backend/src/models/highscore.dto.ts
@@ -8,15 +8,12 @@ export class HighscoreDto {
   score: number;
   @ApiModelProperty()
   seasonId: string;
-  @ApiModelProperty()
-  teamLogotype?: string;
 
   public static from(dto: Partial<HighscoreDto>) {
     const highScoreObj = new HighscoreDto();
     highScoreObj.botName = dto.botName;
     highScoreObj.score = dto.score;
     highScoreObj.seasonId = dto.seasonId;
-    highScoreObj.teamLogotype = dto.teamLogotype ? dto.teamLogotype : null;
     return highScoreObj;
   }
 
@@ -25,17 +22,6 @@ export class HighscoreDto {
       botName: entity.botName,
       score: entity.score,
       seasonId: entity.seasonId,
-    });
-  }
-
-  public static fromRawDataObject(data: {}) {
-    // Used by allBySeasonId in hishscore service
-    // Data includes raw data from highScore table and teams table.
-    return this.from({
-      botName: data["highScores_botName"],
-      score: data["highScores_score"],
-      seasonId: data["highScores_seasonId"],
-      teamLogotype: data["teams_logotypeUrl"],
     });
   }
 }
